Core/Units: moved CharmInfo class and its definitions into a own file
