+++ src/server/game/Spells/Auras/SpellAuraEffects.cpp
@@ -21,6 +21,7 @@
 #include "Battleground.h"
 #include "BattlegroundPackets.h"
 #include "CellImpl.h"
+#include "CharmInfo.h"
 #include "Common.h"
 #include "Containers.h"
 #include "DB2Stores.h"
